Refactoring Settings to move Delete Confirmation modal to separate component
