+++ pages/user/settings.module.scss
@@ -50,11 +50,4 @@
     height: 40px;
     margin-top: $spacing-sm;
   }
-
-  .deleteActions {
-    display: flex;
-    flex-direction: row;
-    gap: $spacing-base;
-    justify-content: center;
-  }
 }
