+++ pages/user/settings.tsx
@@ -1,15 +1,15 @@
-import React, { useState, createRef } from 'react';
+import React, { useState } from 'react';
 import { useTranslation } from 'next-i18next';
 import { useRouter } from 'next/router';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import { useSession, signOut } from 'next-auth/react';
+import { useSession } from 'next-auth/react';
 import { formatDateStringLong, timeElapsed } from 'lib/utils/time';
 import { AppContextProvider } from 'components/App/';
-import ReactMarkdown from 'react-markdown';
 import GlobalHeader from 'components/GlobalHeader';
 import ProfileImage from 'components/User/ProfileImage';
 import Footer from 'components/Footer';
 import Modal from 'components/Modal';
+import ConfirmDelete from 'components/User/Settings/ConfirmDelete';
 
 import type { GetServerSideProps } from 'next';
 
@@ -17,9 +17,7 @@ import styles from './settings.module.scss';
 
 export default function Settings() {
   const { t } = useTranslation();
-  const deleteForm = createRef<HTMLFormElement>();
   const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [userConfirmed, setUserConfirmed] = useState(false);
   const router = useRouter();
 
   const { data: session } = useSession();
@@ -27,26 +25,11 @@ export default function Settings() {
 
   const user = session.user;
 
-  async function destroyAccount() {
-    await fetch('/api/user/account', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ method: 'destroy' })
-    }).then(() => {
-      signOut({ callbackUrl: '/' });
-    });
-  }
-
   function handleAccountDeletion() {
     setShowDeleteConfirm(true);
   }
 
-  function checkUserConfirm(e:React.KeyboardEvent<HTMLInputElement>) {
-    setUserConfirmed((e.currentTarget.value === user.name));
-  }
-
   function cancelDeletion() {
-    deleteForm.current?.reset();
     setShowDeleteConfirm(false);
   }
 
@@ -102,51 +85,8 @@ export default function Settings() {
               { t('Pages.User.Settings.delete_my_account') }
             </button>
 
-            <Modal
-              showModal={showDeleteConfirm}
-              onClose={() => cancelDeletion()}
-            >
-              <div className="container-sm">
-                <h1>{ t('Pages.User.Settings.are_you_sure') }</h1>
-                <p className="text-xl">{ t('Pages.User.Settings.are_you_really_sure') }</p>
-
-                <form autoComplete="off" ref={deleteForm}>
-                  <div className="control">
-                    <label htmlFor="confirmUserName">
-                      <ReactMarkdown>
-                        { t('Pages.User.Settings.enter_username', { user_name: user.name }) }
-                      </ReactMarkdown>
-                    </label>
-
-                    <input
-                      id="confirmUserName"
-                      name="confirmUserName"
-                      type="text"
-                      autoComplete="off"
-                      onKeyUp={checkUserConfirm}
-                    />
-                  </div>
-
-                  <div className={styles.deleteActions}>
-                    <button
-                      type="button"
-                      className="button btn-white"
-                      onClick={cancelDeletion}
-                    >
-                      { t('Pages.User.Settings.cancel') }
-                    </button>
-
-                    <button
-                      type="button"
-                      className="button btn-danger"
-                      onClick={destroyAccount}
-                      disabled={!userConfirmed}
-                    >
-                      { t('Pages.User.Settings.delete') }
-                    </button>
-                  </div>
-                </form>
-              </div>
+            <Modal showModal={showDeleteConfirm} onClose={() => cancelDeletion()}>
+              <ConfirmDelete onCancel={() => cancelDeletion()} />
             </Modal>
           </div>
         </div>
